Dendrite tracing and matching tool
This code was generated on the purpose of analyzing the eGRASP signal of the engram dendrite across multiple days. But this code can be used in general cases. This code can trace the dendrites from the 3D voxel images such as two-photon in vivo z-stack images.  A user friendly interface are implanted for the 3D coordinate tracing. This package also provide the automatic searching of the dendrite from the images taken in different day using the cross-correlation approach. After the dendrite tracing, user not only can see the each dendrite in different days, but also can project the images to dendrite-specified plane.